requirements.txt
colorama==0.4.6
Flask==3.0.3
pipreqs==0.4.7
PyYAML==5.3.1
